Refactor list module tests in core crate
The list module tests were contained in a single test file that was
starting to become too large to properly navigate. This change breaks up
that file into multiple separate test files, based on the general
functionality being tested.
